LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/radio-keyboard-navigation-order-expected.txt










FAIL Radio button keyboard navigation should proceed in tree-order. assert_equals: Navigated to next radio button in form 'inside' expected Element node <input type="radio" name="inside" id="inside1"></input> but got Element node <input id="insideNew" type="radio" form="inside" name="in...
